requirements.txt
# Requirements for Python 3.12+

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning & Forecasting
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0

# Deep Learning (optional, for future TFT implementation)
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
kaggle>=1.5.0

# Notebook & Documentation
jupyter>=1.0.0
ipykernel>=6.22.0
notebook>=6.5.0

# Optional: Experiment Tracking
# wandb>=0.15.0
# mlflow>=2.3.0
# tensorboard>=2.13.0

# Optional: Advanced Features
# simpy>=4.0.0          # Discrete event simulation
# mesa>=2.0.0           # Agent-based modeling
# networkx>=3.1         # Network analysis
# ray[rllib]>=2.5.0     # Reinforcement learning
# gymnasium>=0.28.0     # RL environments
# stable-baselines3>=2.0.0  # RL algorithms

# Optional: Causal Analysis
# dowhy>=0.11.0
# econml>=0.14.0
# pgmpy>=0.1.23

# Statistical Forecasting
neuralforecast>=1.0.0
statsforecast>=1.0.0

# Additional Time Series
gluonts>=0.13.0